Message_Parsing/Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-google-genai
